requirements.txt
absl-py==1.2.0
accelerate==0.13.2
addict==2.4.0
aiohttp==3.8.3
aiosignal==1.2.0
albumentations==1.1.0
altair==4.2.0
antlr4-python3-runtime==4.8
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
basicsr==1.4.2
bcrypt==4.0.1
beautifulsoup4==4.11.1
bitsandbytes==0.35.1
bleach==5.0.1
blinker==1.5
blis==0.7.8
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cmapy==0.6.6
commonmark==0.9.1
confection==0.0.2
contourpy==1.0.5
cryptography==38.0.1
cycler==0.11.0
cymem==2.0.6
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.3.0
diffusers-interpret==0.5.0
dill==0.3.5.1
einops==0.4.1
entrypoints==0.4
executing==1.1.0
facexlib==0.2.5
fastapi==0.85.1
fastjsonschema==2.16.2
ffmpy==0.3.0
filelock==3.8.0
filterpy==1.4.5
fonttools==4.37.4
frozenlist==1.3.1
fsspec==2022.8.2
ftfy==6.1.1
future==0.18.2
-e git+https://github.com/TencentARC/GFPGAN.git@2eac2033893ca7f427f4035d80fe95b92649ac56#egg=gfpgan
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.10.2
google-api-python-client==2.64.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.56.4
gradio==3.7
grpcio==1.49.1
h11==0.12.0
httpcore==0.15.0
httplib2==0.20.4
httpx==0.23.0
huggingface-hub==0.10.0
idna==3.4
imageio==2.14.1
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==7.7.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.10
jsonschema==4.16.0
jupyter-core==4.11.1
jupyter-server==1.19.1
jupyter_client==7.3.5
jupyterlab==3.4.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
jupyterlab_server==2.15.2
kiwisolver==1.4.4
kornia==0.6.0
langcodes==3.3.0
linkify-it-py==1.0.3
llvmlite==0.39.1
lmdb==1.3.0
loguru==0.6.0
lxml==4.9.1
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.1
mdurl==0.1.2
mistune==2.0.4
multidict==6.0.2
murmurhash==1.0.8
nbclassic==0.4.4
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.6.1
nest-asyncio==1.5.6
networkx==2.8.7
notebook==6.4.12
notebook-shim==0.1.0
numba==0.56.3
numpy==1.23.3
oauthlib==3.2.1
omegaconf==2.1.1
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
orjson==3.8.1
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
paramiko==2.11.0
parso==0.8.3
pathy==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
pika==1.3.0
Pillow==9.0.1
preshed==3.0.7
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
ptyprocess==0.7.0
pudb==2019.2
pure-eval==0.2.2
py4j==0.10.9.5
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.9.2
pydeck==0.8.0b3
pyDeprecate==0.3.1
pydub==0.25.1
Pygments==2.13.0
Pympler==1.0.1
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
pyspark==3.3.1
pyTelegramBotAPI==4.8.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-logging-loki==0.3.1
python-multipart==0.0.5
pytorch-lightning==1.5.9
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qudida==0.0.4
realesrgan==0.3.0
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339==6.2
rfc3986==1.5.0
rich==12.5.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.4
stack-data==0.5.1
starlette==0.20.4
streamlit==1.13.0
-e git+https://github.com/CompVis/taming-transformers.git@24268930bf1dce879235a7fddd0b2355b84d7ea6#egg=taming_transformers
tb-nightly==2.11.0a20221026
telebot==0.0.4
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.16.0
test-tube==0.7.5
thinc==8.1.2
threadpoolctl==3.1.0
tifffile==2022.8.12
tinycss2==1.1.1
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.12.1+cu116
torch-fidelity==0.3.0
torchaudio==0.12.1+cu116
torchmetrics==0.6.0
torchvision==0.13.1+cu116
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
transformers==4.22.2
typer==0.4.2
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
uc-micro-py==1.0.1
uritemplate==4.1.1
urllib3==1.26.12
urwid==2.1.2
uvicorn==0.19.0
validators==0.20.0
wasabi==0.10.1
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
websockets==10.4
Werkzeug==2.2.2
widgetsnbextension==3.6.1
yapf==0.32.0
yarl==1.8.1
zipp==3.8.1
